requirements.txt
PyGithub == 1.39